agent: move ENC-related code to a separate module
 #PL-130441
