+++ pkgs/fc/agent/fc/manage/manage.py
@@ -1,20 +1,12 @@
 """Update NixOS system configuration from infrastructure or local sources."""
 
 import argparse
-import filecmp
-import grp
-import hashlib
-import io
-import json
 import os
 import os.path as p
 import re
-import shutil
 import signal
-import socket
 import subprocess
 import sys
-import tempfile
 from datetime import datetime
 from functools import partial
 from pathlib import Path
@@ -24,7 +16,13 @@ import requests
 import structlog
 from fc.maintenance.lib.shellscript import ShellScriptActivity
 from fc.util import nixos
-from fc.util.directory import connect
+from fc.util.enc import (
+    load_enc,
+    seed_enc,
+    update_enc_nixos_config,
+    update_inventory,
+    write_system_state,
+)
 from fc.util.lock import locked
 from fc.util.logging import init_logging
 
@@ -250,181 +248,6 @@ class Channel:
         self.log.info("maintenance-register-succeeded")
 
 
-def load_enc(log, enc_path):
-    """Tries to read enc.json"""
-    global enc
-    try:
-        with open(enc_path) as f:
-            enc = json.load(f)
-    except (OSError, ValueError):
-        # This environment doesn't seem to use an ENC,
-        # i.e. containers. Silently ignore for now.
-        log.info(
-            "no-enc-data",
-            msg="enc data not supported on this infrastructure, ignoring",
-        )
-        enc = {}
-        return
-    return enc
-
-
-def update_enc_nixos_config(log, enc_path):
-    """Update nixos config files managed through the enc."""
-    basedir = os.path.join(os.path.dirname(enc_path), "enc-configs")
-    if not os.path.isdir(basedir):
-        os.makedirs(basedir)
-    previous_files = set(os.listdir(basedir))
-    sudo_srv = grp.getgrnam("sudo-srv").gr_gid
-    for filename, config in enc["parameters"].get("nixos_configs", {}).items():
-        log.info(
-            "update-enc-nixos-config",
-            filename=filename,
-            content=hashlib.sha256(config.encode("utf-8")).hexdigest(),
-        )
-        target = os.path.join(basedir, filename)
-        conditional_update(target, config, encode_json=False)
-        os.chown(target, -1, sudo_srv)
-        previous_files -= {filename}
-    for filename in previous_files:
-        log.info("remove-stale-enc-nixos-config", filename=filename)
-        os.unlink(os.path.join(basedir, filename))
-
-
-def conditional_update(filename, data, encode_json=True):
-    """Updates JSON file on disk only if there is different content."""
-    with tempfile.NamedTemporaryFile(
-        mode="w",
-        suffix=".tmp",
-        prefix=p.basename(filename),
-        dir=p.dirname(filename),
-        delete=False,
-    ) as tf:
-        if encode_json:
-            json.dump(data, tf, ensure_ascii=False, indent=1, sort_keys=True)
-        else:
-            tf.write(data)
-        tf.write("\n")
-        os.chmod(tf.fileno(), 0o640)
-    if not (p.exists(filename)) or not (filecmp.cmp(filename, tf.name)):
-        with open(tf.name, "a") as f:
-            os.fsync(f.fileno())
-        os.rename(tf.name, filename)
-    else:
-        os.unlink(tf.name)
-
-
-def inplace_update(filename, data):
-    """Last-resort JSON update for added robustness.
-
-    If there is no free disk space, `conditional_update` will fail
-    because it is not able to create tempfiles. As an emergency measure,
-    we fall back to rewriting the file in-place.
-    """
-    with open(filename, "r+") as f:
-        f.seek(0)
-        json.dump(data, f, ensure_ascii=False)
-        f.flush()
-        f.truncate()
-        os.fsync(f.fileno())
-
-
-def retrieve(log, directory_lookup, tgt):
-    log.info("retrieve-enc", _replace_msg="Getting: {tgt}", tgt=tgt)
-    try:
-        data = directory_lookup()
-    except Exception:
-        log.error("retrieve-enc-failed", exc_info=True)
-        return
-    try:
-        conditional_update("/etc/nixos/{}".format(tgt), data)
-    except (IOError, OSError):
-        inplace_update("/etc/nixos/{}".format(tgt), data)
-
-
-def write_json(log, calls):
-    """Writes JSON files from a list of (lambda, filename) pairs."""
-    for call in calls:
-        retrieve(log, *call)
-
-
-def system_state(log):
-    def load_system_state():
-        result = {}
-        try:
-            with open("/proc/meminfo") as f:
-                for line in f:
-                    if line.startswith("MemTotal:"):
-                        _, memkb, _ = line.split()
-                        result["memory"] = int(memkb) // 1024
-                        break
-        except IOError:
-            pass
-        try:
-            with open("/proc/cpuinfo") as f:
-                cores = 0
-                for line in f:
-                    if line.startswith("processor"):
-                        cores += 1
-            result["cores"] = cores
-        except IOError:
-            pass
-        return result
-
-    write_json(
-        log,
-        [
-            (lambda: load_system_state(), "system_state.json"),
-        ],
-    )
-
-
-def update_inventory(log):
-    if (
-        not enc
-        or not enc.get("parameters")
-        or not enc["parameters"].get("directory_password")
-    ):
-        log.warning(
-            "update-inventory-no-pass",
-            msg="No directory password. Not updating inventory.",
-        )
-        return
-    try:
-        # For fc-manage all nodes need to talk about *their* environment which
-        # is resource-group specific and requires us to always talk to the
-        # ring 1 API.
-        directory = connect(enc, 1)
-    except socket.error:
-        log.warning(
-            "update-inventory-no-connection",
-            msg="No directory connection. Not updating inventory.",
-        )
-        return
-
-    log.info(
-        "update-inventory",
-        _replace_msg="Getting inventory data from directory...",
-    )
-
-    write_json(
-        log,
-        [
-            (lambda: directory.lookup_node(enc["name"]), "enc.json"),
-            (
-                lambda: directory.list_nodes_addresses(
-                    enc["parameters"]["location"], "srv"
-                ),
-                "addresses_srv.json",
-            ),
-            (lambda: directory.list_permissions(), "permissions.json"),
-            (lambda: directory.list_networks(), "networks.json"),
-            (lambda: directory.list_service_clients(), "service_clients.json"),
-            (lambda: directory.list_services(), "services.json"),
-            (lambda: directory.list_users(), "users.json"),
-        ],
-    )
-
-
 def prepare_switch_in_maintenance(log, build_options, spread, lazy):
     if not enc or not enc.get("parameters"):
         log.warning(
@@ -539,14 +362,6 @@ def maintenance(log, config_file):
     import fc.maintenance.reqmanager
 
     fc.maintenance.reqmanager.transaction(log=log, config_file=config_file)
-
-
-def seed_enc(path):
-    if os.path.exists(path):
-        return
-    if not os.path.exists("/tmp/fc-data/enc.json"):
-        return
-    shutil.move("/tmp/fc-data/enc.json", path)
 
 
 def exit_timeout(log, signum, frame):
@@ -676,6 +491,7 @@ def parse_args():
 
 
 def transaction(log, args):
+    global enc
     seed_enc(args.enc_path)
 
     keep_cmd_output = False
@@ -684,16 +500,16 @@ def transaction(log, args):
     if args.fast:
         build_options.append("--fast")
 
-    load_enc(log, args.enc_path)
+    enc = load_enc(log, args.enc_path)
 
     if args.directory:
-        update_inventory(log)
+        update_inventory(log, enc)
         # reload ENC data in case update_inventory changed something
-        load_enc(log, args.enc_path)
+        enc = load_enc(log, args.enc_path)
         update_enc_nixos_config(log, args.enc_path)
 
     if args.system_state:
-        system_state(log)
+        write_system_state(log)
 
     if args.automatic:
         spread = Spread(
